settings-modal: split lists to components
