+++ client/src/javascript/components/modals/settings-modal/UITab.js
@@ -1,24 +1,22 @@
 import {FormattedMessage, injectIntl} from 'react-intl';
 import React from 'react';
 
-import {Checkbox, Form, FormRow, Select, SelectItem, Radio} from '../../../ui';
-import ErrorIcon from '../../icons/ErrorIcon';
+import {Form, FormRow, Select, SelectItem, Radio} from '../../../ui';
 import Languages from '../../../constants/Languages';
 import ModalFormSectionHeader from '../ModalFormSectionHeader';
 import SettingsStore from '../../../stores/SettingsStore';
 import SettingsTab from './SettingsTab';
-import SortableList from '../../general/SortableList';
-import Tooltip from '../../general/Tooltip';
-import TorrentProperties from '../../../constants/TorrentProperties';
+import TorrentDetailItemsList from './lists/TorrentDetailItemsList';
 
 class UITab extends SettingsTab {
-  tooltipRef = null;
+  constructor(props) {
+    super(props);
 
-  state = {
-    torrentDetails: SettingsStore.getFloodSettings('torrentDetails'),
-    torrentListViewSize: SettingsStore.getFloodSettings('torrentListViewSize'),
-    selectedLanguage: SettingsStore.getFloodSettings('language'),
-  };
+    this.state = {
+      torrentListViewSize: SettingsStore.getFloodSettings('torrentListViewSize'),
+      selectedLanguage: SettingsStore.getFloodSettings('language'),
+    };
+  }
 
   getLanguageSelectOptions() {
     return Object.keys(Languages).map((languageID) => {
@@ -40,29 +38,6 @@ class UITab extends SettingsTab {
     });
   }
 
-  getLockedIDs() {
-    if (this.state.torrentListViewSize === 'expanded') {
-      return ['name', 'eta', 'downloadRate', 'uploadRate'];
-    }
-
-    return [];
-  }
-
-  handleDetailCheckboxValueChange = (id, value) => {
-    let {torrentDetails} = this.state;
-
-    torrentDetails = torrentDetails.map((detail) => {
-      if (detail.id === id) {
-        detail.visible = value;
-      }
-
-      return detail;
-    });
-
-    this.props.onSettingsChange({torrentDetails});
-    this.setState({torrentDetails});
-  };
-
   handleFormChange = ({event, formData}) => {
     if (event.target.type === 'radio') {
       const newState = {torrentListViewSize: formData['ui-torrent-size']};
@@ -79,98 +54,7 @@ class UITab extends SettingsTab {
     }
   };
 
-  handleTorrentDetailsMouseDown = () => {
-    if (this.tooltipRef != null) {
-      this.tooltipRef.dismissTooltip();
-    }
-  };
-
-  handleTorrentDetailsMove = (items) => {
-    this.setState({torrentDetails: items});
-    this.props.onSettingsChange({torrentDetails: items});
-  };
-
-  renderTorrentDetailItem = (item, index) => {
-    const {id, visible} = item;
-    let checkbox = null;
-    let warning = null;
-
-    if (!item.dragIndicator && !this.getLockedIDs().includes(id)) {
-      checkbox = (
-        <span className="sortable-list__content sortable-list__content--secondary">
-          <Checkbox
-            checked={visible}
-            onChange={(event) => this.handleDetailCheckboxValueChange(id, event.target.checked)}
-            modifier="dark">
-            <FormattedMessage id="settings.ui.torrent.details.enabled" />
-          </Checkbox>
-        </span>
-      );
-    }
-
-    if (
-      id === 'tags' &&
-      this.state.torrentListViewSize === 'expanded' &&
-      index < this.state.torrentDetails.length - 1
-    ) {
-      const tooltipContent = <FormattedMessage id="settings.ui.torrent.details.tags.placement" />;
-
-      warning = (
-        <Tooltip
-          className="tooltip tooltip--is-error"
-          content={tooltipContent}
-          offset={-5}
-          ref={(ref) => {
-            this.tooltipRef = ref;
-          }}
-          scrollContainer={this.props.scrollContainer}
-          width={200}
-          wrapperClassName="sortable-list__content sortable-list__content--secondary tooltip__wrapper"
-          wrapText>
-          <ErrorIcon />
-        </Tooltip>
-      );
-    }
-
-    const content = (
-      <div className="sortable-list__content sortable-list__content__wrapper">
-        {warning}
-        <span className="sortable-list__content sortable-list__content--primary">
-          <FormattedMessage id={TorrentProperties[id].id} />
-        </span>
-        {checkbox}
-      </div>
-    );
-
-    if (item.dragIndicator) {
-      return <div className="sortable-list__item">{content}</div>;
-    }
-
-    return content;
-  };
-
   render() {
-    const lockedIDs = this.getLockedIDs();
-    let torrentDetailItems = this.state.torrentDetails.slice();
-
-    if (this.state.torrentListViewSize === 'expanded') {
-      let nextUnlockedIndex = lockedIDs.length;
-
-      torrentDetailItems = torrentDetailItems
-        .reduce((accumulator, detail) => {
-          const lockedIDIndex = lockedIDs.indexOf(detail.id);
-
-          if (lockedIDIndex > -1) {
-            accumulator[lockedIDIndex] = detail;
-          } else {
-            accumulator[nextUnlockedIndex++] = detail;
-          }
-
-          return accumulator;
-        }, [])
-        .filter((item) => item != null);
-    }
-
     return (
       <Form onChange={this.handleFormChange}>
         <ModalFormSectionHeader>
@@ -207,13 +91,9 @@ class UITab extends SettingsTab {
           <FormattedMessage id="settings.ui.displayed.details" />
         </ModalFormSectionHeader>
         <FormRow>
-          <SortableList
-            className="sortable-list--torrent-details"
-            items={torrentDetailItems}
-            lockedIDs={lockedIDs}
-            onMouseDown={this.handleTorrentDetailsMouseDown}
-            onDrop={this.handleTorrentDetailsMove}
-            renderItem={this.renderTorrentDetailItem}
+          <TorrentDetailItemsList
+            torrentListViewSize={this.state.torrentListViewSize}
+            onSettingsChange={this.props.onSettingsChange}
           />
         </FormRow>
       </Form>
